refactor(structure): app/routes — kebab folders + PascalCase route components
Per the project-structure skill, route folders are the only place
kebab-case is allowed (URL-shaped). Routes were flat kebab .tsx files
at app/routes/; promote each to a folder with a barrel index.tsx and
the component file inside as PascalCase.

Layout changes:

- 17 routes promoted: accept-invite/, access-required/, invitation-view/,
  invite-speaker/, login/, notifications/, plan-prayers/, plan-speakers/,
  prepare-invitation/, prepare-prayer-invitation/, profile/, templates/,
  templates-prayer-letter/, templates-programs/, templates-speaker-letter/,
  ward-settings/, week/

- invite-speaker/'s 5 sub-components (chat drawer, CTA banner, session
  gate, share toolbar, states) move INTO that folder and rename to
  PascalCase (SpeakerChatFloatingDrawer, SpeakerChatCTABanner, SessionGate,
  ShareToolbar, InviteSpeakerStates) — they're route-specific, not
  shared.

- prepare-invitation/'s sub-components (PrepareInvitationHeader) and
  validation util (prepare-invitation-validation.ts → utils/
  prepareInvitationValidation.ts) move into that route's folder.

- prepare-prayer-invitation/'s PreparePrayerInvitationHeader moves into
  that route's folder.

- PrepareInvitationPageMessage.tsx stays at routes/ root — it's shared
  between the two prepare-* routes.

External @/app/routes/<route> imports survive unchanged thanks to the
barrel index.tsx. Internal sibling imports rewritten to match the new
layout. Typecheck + 346 tests + build all pass.
