+++ src/app/routes/accept-invite/AcceptInvitePage.tsx
