+++ src/app/routes/accept-invite/index.tsx
@@ -1,0 +1,1 @@
+export * from "./AcceptInvitePage";
