+++ src/app/routes/access-required/index.tsx
@@ -1,0 +1,1 @@
+export * from "./AccessRequired";
