+++ src/app/routes/invitation-view/index.tsx
@@ -1,0 +1,1 @@
+export * from "./InvitationViewPage";
