+++ src/app/routes/invite-speaker/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SpeakerInvitationLandingPage";
