+++ src/app/routes/login/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Login";
