+++ src/app/routes/notifications/index.tsx
@@ -1,0 +1,1 @@
+export * from "./NotificationsPage";
