+++ src/app/routes/plan-prayers/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PlanPrayersRoute";
