+++ src/app/routes/plan-speakers/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PlanSpeakersRoute";
