+++ src/app/routes/prepare-invitation/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PrepareInvitationPage";
