+++ src/app/routes/prepare-prayer-invitation/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PreparePrayerInvitationPage";
