+++ src/app/routes/profile/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ProfilePage";
