+++ src/app/routes/templates-prayer-letter/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PrayerLetterTemplatePage";
