+++ src/app/routes/templates-programs/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ProgramTemplatesPage";
