+++ src/app/routes/templates-speaker-letter/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SpeakerLetterTemplatePage";
