+++ src/app/routes/templates/index.tsx
@@ -1,0 +1,1 @@
+export * from "./TemplatesPage";
