+++ src/app/routes/ward-settings/index.tsx
@@ -1,0 +1,1 @@
+export * from "./WardSettingsPage";
