+++ src/app/routes/week/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Week";
